+++ src/main/java/lebron/DeleteCommand.java
@@ -1,3 +1,5 @@
+package lebron;
+
 public class DeleteCommand extends Command {
 
     private int index;
