+++ src/main/java/lebron/Event.java
@@ -1,3 +1,5 @@
+package lebron;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
