+++ src/main/java/lebron/ExitCommand.java
@@ -1,3 +1,4 @@
+package lebron;
 
 public class ExitCommand extends Command {
