+++ src/main/java/lebron/LeBronException.java
@@ -1,3 +1,5 @@
+package lebron;
+
 public class LeBronException extends Exception {
     
     public LeBronException(String message) {
