+++ src/main/java/lebron/MarkCommand.java
@@ -1,16 +1,18 @@
-public class UnmarkCommand extends Command {
+package lebron;
 
+public class MarkCommand extends Command {
+    
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws LeBronException {
         Task task = taskList.getTask(index);
-        taskList.unmarkTask(index);
-        ui.showTaskUnmarked(task);
+        taskList.markTask(index);
+        ui.showTaskMarked(task);
         storage.saveTasks(taskList);
     }
